Added two examples, inbuilt_interactive_mode.py with argparse for unity build and 2nd was renamed bounding_boxes file. Removed old functions in file and use gym_task_utils. Added line option from bounding boxes to crosshair and printed distance
